requirements.txt
streamlit>=1.32.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.18.0
ta>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
pytz>=2024.1
streamlit-autorefresh>=1.0.1
lxml>=5.1.0
groq>=0.9.0
google-genai>=1.65.0
python-dotenv>=1.0.0
urllib3>=2.0.0
